fix validator and params handler classes

+++ src/Console/Commands/Compound/DeleteNodeCommand.php
@@ -2,21 +2,32 @@
 
 namespace LKSS\Console\Commands\Compound;
 
+use LKSS\Console\Commands\AbstractClasses\AbstractCompoundCommand;
+use LKSS\Console\Commands\CommandParamsHandler;
+use LKSS\Console\Commands\Validation\CommandValidator;
+use LKSS\Console\Commands\Validation\Rules\DeleteNodeRule;
 use LKSS\StorageTree;
-use LKSS\Console\Commands\Interfaces\CompoundCommand;
 
-class DeleteNodeCommand implements CompoundCommand
+class DeleteNodeCommand extends AbstractCompoundCommand
 {
     private StorageTree $storage;
 
     public function __construct(StorageTree $storage)
     {
         $this->storage = $storage;
+        $rule = new DeleteNodeRule();
+        $this->validator = new CommandValidator($rule, self::DELETE_NODE);
+        $this->paramsHandler = new CommandParamsHandler($rule, self::DELETE_NODE);
     }
 
     public function execute(string $command): void
     {
-        $nodeKey = substr($command, strlen(self::DELETE_NODE) + 1);
-        $this->storage->deleteNode($nodeKey);
+        if ($this->validator->isValid($command)) {
+            $this->paramsHandler->setCommand($command);
+            $nodeKey = $this->paramsHandler->getCurrentParam();
+            $this->storage->deleteNode($nodeKey);
+        } else {
+            echo "Dont valid params!\n";
+        }
     }
 }
